services/simulator/requirements.txt
kafka-python==2.0.2
pydantic==2.11.1
